requirements.txt
pandas
numpy
scikit-learn
streamlit
spotipy
python-dotenv
